LPS-121610 Creates a CSSVariablesProvider for default Theme token values
To avoid having to hardcode default token values in the css implementation
and simplify the generation of new tokens, this CSSVariablesProvider
provides default values for all tokens in a token definition with a
cssVariable mapping.

The variable definitions are pushed to the ":root" element so they apply
to the whole page and can be overridden in context using css scopes.