LPS-121753 Fix test failure in remote-app-client-js
Tests were dying with:

    TypeError: _index.__LIFERAY_REMOTE_APP_SDK__.Client is not a constructor

For much the same reason as we saw and fixed in "LPS-121753 Don't use
object concise methods for AUI component constructors" some commits
back; we can't use object concise methods as constructor functions, and
we were previously able to do it only because our Babel config turned
them into something other than concise functions.

So this commit fixes that by hoisting the concise method out to the
top level (which is nice in any case because it reduces indentation
by one level), but note that even though that is enough to unbreak
the tests, I am *also* applying the same kind of config that I did in
analytics-client-js, I don't think we officially want the SDK to drop
IE support (seeing as it was "included" in 7.3, even though we still
haven't published it to the CDN).