LPS-124403 Add css class for the background color too in cases where it may be a class (if start with # is considered a hex value so it is not added)
Due to the change in common styles we changed the way background colors were used, instead of adding a class like `bg-` a hex value or a token is now used. For upgrading what we did is trying to convert from class to its color value but as the classes are customizable we cannot convert everything. For that cases just preserve the old behaviour of adding a class to the element prefixing with `bg-`