LPS-129456 Update inputPlaceHolder when editor is initialized
When the browser's viewport is resized, we [destroy and recreate the
editor](https://git.io/JYLr3), to update it's toolbars.

When this editor is referenced by an `input_localized` component,
this will cause the issue mentioned in [LPS-129456](https://issues.liferay.com/browse/LPS-129456)

To fix the issue, we make use of the `InputEditorTag`'s `onInit` attribute
which will call a new `updateInputPlaceholder` function defined in
`input_localized` to update its `inputPlaceholder` property if needed.