LPS-130651 users-admin-impl: fixes UserSearchPermissionFilterContributor
This commit does two things:

- Removes logic: no longer includes all users by default. This should fix the
performance concerns described in the ticket:
https://issues.liferay.com/browse/LPS-130651

Before, all users were being included by default, which meant that all the filtering
was done in the permission post-filter. In the case where a user does not have
permissions to view any users, this was very slow.

With this change, pre-filtering is no longer clobbered, meaning far fewer
results are returned.

- Adds logic: If a user has MANAGE_USERS permission for an organization, include that organization's users